backend/requirements.txt
flask
python-dotenv
google-generativeai
algosdk
flask-cors
